refactor: extract option utilities
