+++ lib/utils/normalizeOptions.js
@@ -1,33 +1,4 @@
-/**
- * Sets a constant default value for the property when it is undefined.
- * @template T
- * @template {keyof T} Property
- * @param {T} object An object.
- * @param {Property} property A property of the provided object.
- * @param {T[Property]} defaultValue The default value to set for the property.
- * @returns {T[Property]} The defaulted property value.
- */
-const d = (object, property, defaultValue) => {
-  if (typeof object[property] === 'undefined' && typeof defaultValue !== 'undefined') {
-    object[property] = defaultValue;
-  }
-  return object[property];
-};
-
-/**
- * Resolves the value for a nested object option.
- * @template T
- * @template {keyof T} Property
- * @template Result
- * @param {T} object An object.
- * @param {Property} property A property of the provided object.
- * @param {function(T | undefined): Result} fn The handler to resolve the property's value.
- * @returns {Result} The resolved option value.
- */
-const nestedOption = (object, property, fn) => {
-  object[property] = fn(object[property]);
-  return object[property];
-};
+const { d, n } = require('../../options');
 
 /**
  * Normalizes the options for the plugin.
@@ -40,7 +11,7 @@ const normalizeOptions = (options) => {
   d(options, 'forceEnable');
   d(options, 'library');
 
-  nestedOption(options, 'overlay', (overlay) => {
+  n(options, 'overlay', (overlay) => {
     /** @type {import('../types').NormalizedErrorOverlayOptions} */
     const defaults = {
       entry: require.resolve('../../client/ErrorOverlayEntry'),
